% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arglist.R
\name{arglist}
\alias{arglist}
\title{An argument list}
\usage{
arglist(...)
}
\arguments{
\item{...}{objects, including \code{\link{promise}}s, possibly with names.}
}
\value{
a \code{\link{list}} with class \code{c("uneval_arglist", "list")}.
}
\description{
A lightweight wrapper around \code{\link{list}} designed to store values to be passed
to function arguments, including \code{\link{promise}}s.
}
\details{
\code{arglist}s are \code{\link{list}}s that have more useful output for any \code{\link{promise}}s
they contains when \code{\link{print}}ed or \code{\link{format}}ed, as \code{\link{promise}}s by themselves
do not print well. They otherwise act as normal \code{\link{list}}s.
}
\examples{
# promises in normal lists don't print well
list(promise(x + 1))

# arglists show more useful information
arglist(promise(x + 1))

}
\concept{promise list constructors and predicates}
