% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopartial.R
\name{capture_all}
\alias{capture_all}
\alias{capture_named}
\alias{capture_dots}
\title{Capture promises for arguments to a function call}
\usage{
capture_all(which = sys.parent())

capture_named(which = sys.parent())

capture_dots(which = sys.parent())
}
\arguments{
\item{which}{(scalar \link{integer}) the frame number to get call information
from, as returned by a function like \code{\link[=sys.parent]{sys.parent()}}. The default looks at
the arguments passed to the function that called this one.}
}
\value{
A \link{promise_list} where names are names of arguments to the function
in the given frame and values are the promises corresponding to those
arguments.
}
\description{
Capture a list of \link{promise}s representing arguments to the surrounding
function (or another function specified by \code{which}).
}
\details{
\code{capture_all()} captures all arguments to the function call.

\code{capture_named()} captures named arguments (i.e. those explicitly
listed in the function definition).

\code{capture_dots()} captures arguments passed via \code{...}
}
\examples{
f = function(x, y, ...) capture_all()
f(1, y = 2, z = 3)

f = function(x, y, ...) capture_named()
f(1, y = 2, z = 3)

f = function(x, y, ...) capture_dots()
f(1, y = 2, z = 3)

}
