% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{DEFAULT}
\alias{DEFAULT}
\title{A default argument}
\usage{
DEFAULT()
}
\description{
A flag indicating that the default value of an argument should be used.
}
\details{
\code{\link[=DEFAULT]{DEFAULT()}} is a flag passed to an argument of an \link{autopartial} function
that indicates the argument should keep its default value (or the most
recently partially-applied value of that argument). It is styled in all caps,
like \code{NULL} or \code{TRUE}, as a reminder that it is a special flag.
}
\examples{
f = autopartial(function(x, y = "b") {
  c(x = x, y = y)
})

f("a")

# uses the default value of `y` ("b")
f("a", y = DEFAULT())

# partially apply `f`
g = f(y = "c")
g

# uses the last partially-applied value of `y` ("c")
g("a", y = DEFAULT())
}
\seealso{
\code{\link[=autopartial]{autopartial()}}

Other DEFAULT constructors and predicates: 
\code{\link{is_default}()}
}
\concept{DEFAULT constructors and predicates}
