% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise_list.R
\name{is_promise_list}
\alias{is_promise_list}
\title{Is \code{x} a promise list?}
\usage{
is_promise_list(x)
}
\arguments{
\item{x}{an object.}
}
\value{
scalar \link{logical}.
}
\description{
\code{TRUE} if \code{x} is a \link{promise_list}, \code{FALSE} otherwise.
}
\examples{
is_promise_list(promise_list())

}
\seealso{
Other promise list constructors and predicates: 
\code{\link{promise_list}()}
}
\concept{promise list constructors and predicates}
