% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiver.R
\name{is_waiver}
\alias{is_waiver}
\title{Is \code{x} a waiver?}
\usage{
is_waiver(x)
}
\arguments{
\item{x}{an object.}
}
\value{
scalar \code{\link{logical}}.
}
\description{
\code{TRUE} if \code{x} is a \code{\link{waiver}} or a promise to a \code{\link{waiver}}, \code{FALSE}
otherwise. Designed to detect waivers in \code{\link{arglist}}s without evaluating
arguments in many cases.
}
\examples{
is_waiver(waiver())
is_waiver(promise(waiver()))
x = waiver()
is_waiver(promise(x))

}
\seealso{
Other waiver constructors and predicates: 
\code{\link{waiver}()}
}
\concept{waiver constructors and predicates}
