% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopartial.R
\name{new_autopartial}
\alias{new_autopartial}
\title{Low-level constructor for automatically partially-applied functions}
\usage{
new_autopartial(
  f,
  args = promise_list(),
  ...,
  required_arg_names = find_required_arg_names(f),
  waivable = TRUE,
  f_expr = substitute(f)
)
}
\arguments{
\item{f}{(\link{closure} or \link{primitive}) function to automatically partially-apply}

\item{args}{(\link{list}; typically a \link{promise_list}) arguments to apply now}

\item{...}{ignored.}

\item{required_arg_names}{(\link{character} vector) names of required arguments
in \code{f}. When all of these have been supplied, the function will be evaluated.
The default, \code{find_required_arg_names(f)}, considers all arguments without a
default value in the function definition to be required. Pass \code{NULL} or
\code{character()} to get tradition (non-automatic) partial application.}

\item{waivable}{(scalar \link{logical}) if \code{TRUE}, if you pass \code{waiver()} to an
argument to this function, whatever value that argument already has will be
used instead.}

\item{f_expr}{(\link{language}) an expression representing \code{f}. Used
primarily for cosmetic purposes, such as printing, and to construct values
given to \code{\link[=sys.call]{sys.call()}} in the underlying function.}
}
\value{
a \link{function} that when called will be partially applied until all of
the arguments in \code{required_arg_names} have been supplied.
}
\description{
Construct a version of the function \code{f} that is partially applied when called
unless all required arguments have been supplied. This is a low-level
constructor that should be used only if you need to manually adjust
\code{required_arg_names}, \code{waivable}, or \code{f_expr}. In most cases, you should use
the higher-level interfaces \code{\link[=autopartial]{autopartial()}} or \code{\link[=partial]{partial()}}.
}
\examples{
# TODO

}
