% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\name{promise_env}
\alias{promise_env}
\title{Get a promise environment}
\usage{
promise_env(x)
}
\arguments{
\item{x}{<\code{\link{promise}}>}
}
\value{
an \code{\link{environment}} if \code{x} is a \code{\link{promise}}; otherwise \code{NULL}.
}
\description{
Return the \code{\link{environment}} associated with a \code{\link{promise}}.
}
\examples{
promise_env(promise(x + 1))

}
