% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\name{promise_expr}
\alias{promise_expr}
\title{Get a promise expression}
\usage{
promise_expr(x)
}
\arguments{
\item{x}{<\code{\link{promise}}>}
}
\value{
an expression if \code{x} is a \code{\link{promise}}; otherwise \code{x}.
}
\description{
Return the expression associated with a \code{\link{promise}}.
}
\examples{
promise_expr(promise(x + 1))

}
