% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise_list.R
\name{promise_list}
\alias{promise_list}
\title{A promise list}
\usage{
promise_list(...)
}
\arguments{
\item{...}{\link{promise}s (or other objects), possibly with names.}
}
\value{
a \link{list} with class \code{c("uneval_promise_list", "list")}.
}
\description{
A lightweight wrapper around \link{list} designed to store
\link{promise}s. It is allowed to contain other types of objects as well.
}
\details{
The primary use of \link{promise_list}s are as \link{list}s that have more useful
output for any \link{promise}s they contains when \link{print}ed or \link{format}ed, as
\link{promise}s by themselves do not print well.
}
\examples{
# promises in normal lists don't print well
list(promise(x + 1))

# promise_lists show more useful information
promise_list(promise(x + 1))

}
\seealso{
Other promise list constructors and predicates: 
\code{\link{is_promise_list}()}
}
\concept{promise list constructors and predicates}
