% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\name{promises}
\alias{promises}
\title{Construct a list of promises}
\usage{
promises(...)
}
\arguments{
\item{...}{bare expressions to wrap in promises.}
}
\value{
an \code{\link{arglist}} of \code{\link{promise}}s.
}
\description{
Create an \code{\link{arglist}} of promises from bare expressions.
}
\examples{
promises(1, x + 1, y)

}
\seealso{
\code{\link[=promise_expr]{promise_expr()}} and \code{\link[=promise_env]{promise_env()}} for extracting the expression
or environment associated with a \code{\link{promise}}.

\code{\link[=capture]{capture()}} for capturing the \code{\link{promise}} associated with a single
function argument.

\code{\link[=capture_all]{capture_all()}}, \code{\link[=capture_named]{capture_named()}}, and \code{\link[=capture_dots]{capture_dots()}} for
capturing the \code{\link{promise}}s associated with multiple function arguments.

Other promise constructors: 
\code{\link{as_promise}()},
\code{\link{new_promise}()},
\code{\link{promise}()}
}
\concept{promise constructors}
