% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_partial.R
\name{waiver}
\alias{waiver}
\title{A waived argument}
\usage{
waiver()
}
\description{
A flag indicating that the default value of an argument should be used.
}
\details{
A \code{\link[=waiver]{waiver()}} is a flag passed to an argument of an \link{auto_partial} function
that indicates the arguments should keep its default value (or the most
recently partially-applied value of that argument).
}
\examples{
f = auto_partial(function(x, y = "b") {
  c(x = x, y = y)
})

f("a")

# uses the default value of `y` ("b")
f("a", y = waiver())

# partially apply `f`
g = f(y = "c")
g

# uses the last partially-applied value of `y` ("c")
g("a", y = waiver())
}
\seealso{
\code{\link[=auto_partial]{auto_partial()}}
}
