% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopartial.R
\name{new_autopartial}
\alias{new_autopartial}
\title{Low-level constructor for automatically partially-applied functions}
\usage{
new_autopartial(
  f,
  args = arglist(),
  required_arg_names = find_required_arg_names(f),
  waivable = TRUE,
  f_expr = substitute(f)
)
}
\arguments{
\item{f}{<\code{\link{closure}} | \code{\link{primitive}}> function to automatically partially-apply.}

\item{args}{<\code{\link{list}} | \code{\link{pairlist}} | \code{\link{arglist}}> arguments to apply now.}

\item{required_arg_names}{<\code{\link{character}}> names of required arguments
in \code{f}. When all of these have been supplied, the function will be evaluated.
The default, \code{find_required_arg_names(f)}, considers all arguments without a
default value in the function definition to be required. Pass \code{NULL} or
\code{character()} to get traditional (non-automatic) partial application.}

\item{waivable}{<\code{\link{logical}}> if \code{TRUE}, if you pass \code{waiver()} to an
argument to this function, whatever value that argument already has will be
used instead.}

\item{f_expr}{<\code{\link{language}}> an expression representing \code{f}. This is primarly
cosmetic; it does not affect what function is called, but may be used for
printing purposes and will be available as the first element of \code{\link[=sys.call]{sys.call()}}
within \code{f}.}
}
\value{
a \code{\link{function}} that when called will be partially applied until all
of the arguments in \code{required_arg_names} have been supplied.
}
\description{
Construct a version of the function \code{f} that is partially applied when called
unless all required arguments have been supplied. This is a low-level
constructor that should be used only if you need to manually adjust
\code{required_arg_names}, \code{waivable}, or \code{f_expr}. In most cases, you should use
the higher-level interfaces \code{\link[=autopartial]{autopartial()}} or \code{\link[=partial]{partial()}}.
}
\examples{
# TODO

}
\seealso{
\code{\link[=autopartial]{autopartial()}} and \code{\link[=partial]{partial()}} for higher-level interfaces to
constructing partial functions.
}
